# docs/spec/openapi.yaml
openapi: 3.1.0
info:
  title: BlindSight Stage 0/1 API
  version: 0.1.0
  description: >-
    Text-only captured-view understanding API. Clients own camera interaction, speech, and
    earcons. The API makes no safety, navigation, or mobility-aid claim.
servers:
  - url: https://example--blindsight-web.modal.run
    description: Deployed Modal application; replace with the published deployment URL.
  - url: https://example.trycloudflare.com
    description: Local development through a Cloudflare quick tunnel.
security:
  - ApiKey: []
...
paths:
  /v1/excerpts:
    get:
      tags: [Excerpts]
      operationId: listExcerpts
      summary: List preloaded demonstration excerpts
      responses:
        "200":
          description: Available excerpts
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Excerpt"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/excerpts/{excerpt_id}/poster:
    parameters:
      - $ref: "#/components/parameters/ExcerptId"
    get:
      tags: [Excerpts]
      operationId: getExcerptPoster
      summary: Get the still poster for one demonstration excerpt
      description: >-
        Requires the X-API-Key header like every other /v1 operation. Clients fetch the poster
        programmatically and render it from an object URL; a bare image source cannot send the
        header.
      responses:
        "200":
          description: JPEG poster frame
          headers:
            Cache-Control:
              schema:
                type: string
              example: public, max-age=86400
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/captures:
    post:
      tags: [Captures]
      operationId: createCapture
      summary: Start a fresh capture and scene session
      description: >-
        A live source begins in recording state. An excerpt source begins processing immediately.
        Every call creates a fresh scene session; captured evidence is never merged with a prior
        session. Ending an earlier scene session is a client action: the shared key carries no
        client identity, so the server cannot infer which session a new capture supersedes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaptureRequest"
            examples:
              live:
                value:
                  source:
                    type: live
                    mime_type: video/webm
              excerpt:
                value:
                  source:
                    type: excerpt
                    excerpt_id: via-001-entry-02
      responses:
        "201":
          description: Capture created
          headers:
            Location:
              description: Poll URL for the capture resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureResource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMedia"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/captures/{capture_id}/chunks/{index}:
    parameters:
      - $ref: "#/components/parameters/CaptureId"
      - name: index
        in: path
        required: true
        description: Zero-based chunk index. Arrival order is irrelevant.
        schema:
          type: integer
          minimum: 0
          maximum: 9999
    put:
      tags: [Captures]
      operationId: putCaptureChunk
      summary: Upload one numbered live-capture chunk
      description: >-
        Repeating an index with byte-identical content is idempotent. Repeating an index with
        different content returns CHUNK_CONFLICT. Chunks are accepted only while status is
        recording.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Chunk accepted or confirmed idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChunkReceipt"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/TooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMedia"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/captures/{capture_id}/complete:
    parameters:
      - $ref: "#/components/parameters/CaptureId"
    post:
      tags: [Captures]
      operationId: completeCapture
      summary: Close live ingestion and start asynchronous understanding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteCaptureRequest"
            example:
              chunk_count: 8
              mime_type: video/webm
      responses:
        "202":
          description: Capture accepted for processing
          headers:
            Location:
              schema:
                type: string
            Retry-After:
              schema:
                type: integer
                minimum: 1
              example: 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureResource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "415":
          $ref: "#/components/responses/UnsupportedMedia"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/captures/{capture_id}:
    parameters:
      - $ref: "#/components/parameters/CaptureId"
    get:
      tags: [Captures]
      operationId: getCapture
      summary: Poll capture state and retrieve the scene card
      responses:
        "200":
          description: Current capture resource
          headers:
            Retry-After:
              description: Present while processing
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptureResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/scene-sessions/{scene_session_id}/questions:
    parameters:
      - $ref: "#/components/parameters/SceneSessionId"
    post:
      tags: [Questions]
      operationId: createQuestion
      summary: Ask the active scene card first
      description: >-
        This operation never invokes the video provider. A card miss settles as
        needs_clip_consent so the client can obtain explicit user consent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionRequest"
            example:
              question: What colour was the mug on the table?
      responses:
        "202":
          description: Card-grounded question accepted
          headers:
            Location:
              schema:
                type: string
            Retry-After:
              schema:
                type: integer
                minimum: 1
              example: 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionResource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/scene-sessions/{scene_session_id}/questions/{question_id}:
    parameters:
      - $ref: "#/components/parameters/SceneSessionId"
      - $ref: "#/components/parameters/QuestionId"
    get:
      tags: [Questions]
      operationId: getQuestion
      summary: Poll a card or captured-view question
      responses:
        "200":
          description: Current question resource
          headers:
            Retry-After:
              description: Present while processing
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/scene-sessions/{scene_session_id}/questions/{question_id}/clip-check:
    parameters:
      - $ref: "#/components/parameters/SceneSessionId"
      - $ref: "#/components/parameters/QuestionId"
    post:
      tags: [Questions]
      operationId: checkCapturedView
      summary: Re-check the stored captured view after explicit consent
      description: >-
        Valid only when the question is in needs_clip_consent. Calling this operation is the
        client's assertion that explicit user consent was obtained.
      responses:
        "202":
          description: Captured-view check accepted
          headers:
            Location:
              schema:
                type: string
            Retry-After:
              schema:
                type: integer
                minimum: 1
              example: 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/scene-sessions/{scene_session_id}:
    parameters:
      - $ref: "#/components/parameters/SceneSessionId"
    delete:
      tags: [Questions]
      operationId: deleteSceneSession
      summary: End the scene session and clear its conversation
      description: >-
        Ends the conversation for this scene session. Later question operations on it return
        NOT_FOUND. The capture resource remains readable and retained evidence is unaffected.
      responses:
        "204":
          description: Scene session ended
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/transition-sessions:
    post:
      tags: [Transitions]
      operationId: createTransitionSession
      summary: Create and enable one continuous transition session
      description: >-
        Continuous transition processing is off until this authenticated operation creates a
        session. A session reports only transition offers; it never creates a capture or a scene card.
      responses:
        "201":
          description: Transition session created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionSessionResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/transition-sessions/{transition_session_id}/chunks/{index}:
    parameters:
      - $ref: "#/components/parameters/TransitionSessionId"
      - $ref: "#/components/parameters/TransitionChunkIndex"
    put:
      tags: [Transitions]
      operationId: putTransitionChunk
      summary: Add one numbered continuous-media chunk
      description: >-
        Chunks can arrive out of order. Only the contiguous prefix is processed. The Content-Type
        header declares the media type. Repeating identical bytes at an index is idempotent.
      requestBody:
        required: true
        content:
          video/webm:
            schema:
              type: string
              format: binary
          video/mp4:
            schema:
              type: string
              format: binary
          video/quicktime:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Chunk stored or accepted idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionChunkReceipt"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/TransitionQueueTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMedia"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/transition-sessions/{transition_session_id}:
    parameters:
      - $ref: "#/components/parameters/TransitionSessionId"
    get:
      tags: [Transitions]
      operationId: getTransitionSession
      summary: Read transition-session state and retained transition events
      parameters:
        - name: cursor
          in: query
          required: false
          description: >-
            Return only events after this retained transition-event identifier. Reusing a cursor
            returns the same ordered events until newer events are retained.
          schema:
            oneOf:
              - $ref: "#/components/schemas/Identifier"
              - type: "null"
      responses:
        "200":
          description: Current transition-session resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransitionSessionResource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Transitions]
      operationId: deleteTransitionSession
      summary: Stop a transition session and release avoidable work
      responses:
        "204":
          description: Stop request accepted; later media ingestion is rejected.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Shared team key stored in a Modal secret.
  parameters:
    CaptureId:
      name: capture_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    ExcerptId:
      name: excerpt_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
...
  responses:
    BadRequest:
      description: Malformed or semantically invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: INVALID_REQUEST
              message: chunk_count must be at least 1
              retryable: false
              details: {}
    Unauthorized:
      description: Missing or incorrect shared API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: UNAUTHORIZED
              message: A valid X-API-Key header is required.
              retryable: false
              details: {}
    NotFound:
      description: Requested resource does not exist or is no longer active
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: NOT_FOUND
              message: The requested resource was not found.
              retryable: false
              details: {}
...
    TooLarge:
      description: Chunk or accumulated capture exceeds configured limits
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: CAPTURE_TOO_LARGE
              message: The capture exceeds the configured byte limit.
              retryable: false
              details: {}
...
    UnsupportedMedia:
      description: Unsupported capture MIME type
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: UNSUPPORTED_MEDIA_TYPE
              message: Supported live capture types are video/webm, video/mp4, and video/quicktime.
              retryable: false
              details: {}
    InternalError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: INTERNAL_ERROR
              message: The request could not be completed.
              retryable: true
              details: {}
  schemas:
...
    Excerpt:
      type: object
      additionalProperties: false
      required: [excerpt_id, label, duration_seconds, poster_url]
      properties:
        excerpt_id:
          $ref: "#/components/schemas/Identifier"
        label:
          type: string
          minLength: 1
          example: Elevator lobby
        duration_seconds:
          type: number
          minimum: 0.1
          maximum: 30
          example: 8
        poster_url:
          type: string
          format: uri-reference
          example: /v1/excerpts/via-001-entry-02/poster
...
    CreateCaptureRequest:
      type: object
      additionalProperties: false
      required: [source]
      properties:
        source:
          $ref: "#/components/schemas/CaptureSource"
    CompleteCaptureRequest:
      type: object
      additionalProperties: false
      required: [chunk_count, mime_type]
      properties:
        chunk_count:
          type: integer
          minimum: 1
          maximum: 10000
          description: Declared total; every index from zero through chunk_count minus one must exist.
        mime_type:
          type: string
          enum: [video/webm, video/mp4, video/quicktime]
    ChunkReceipt:
      type: object
      additionalProperties: false
      required: [capture_id, index, bytes, idempotent]
      properties:
        capture_id:
          $ref: "#/components/schemas/Identifier"
        index:
          type: integer
          minimum: 0
        bytes:
          type: integer
          minimum: 1
        idempotent:
          type: boolean
          description: True when identical bytes were already stored at this index.
...
    CaptureResource:
      type: object
      additionalProperties: false
      required:
        - capture_id
        - scene_session_id
        - source
        - status
        - card
        - failure
        - created_at
        - updated_at
      properties:
        capture_id:
          $ref: "#/components/schemas/Identifier"
        scene_session_id:
          $ref: "#/components/schemas/Identifier"
        source:
          $ref: "#/components/schemas/CaptureSource"
        status:
          $ref: "#/components/schemas/CaptureStatus"
        card:
          oneOf:
            - $ref: "#/components/schemas/SceneCard"
            - type: "null"
        failure:
          oneOf:
            - $ref: "#/components/schemas/AsyncFailure"
            - type: "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...